Add worm image rotation and rework body collision

+++ Harvester.pde
@@ -1,5 +1,5 @@
 class Harvester {
-  int x, y; // Position on screen
+  float x, y; // Position on screen
   int speed = 5; // Speed that the harvester will move if the gamemode is appropriate
   
   Harvester() {
